pyproject.toml
[tool.poetry]
name = "plainx-sentry"
packages = [
    { include = "plainx" },
]

version = "0.1.0"
description = ""
authors = ["Dave Gaeddert <dave.gaeddert@dropseed.dev>"]

[tool.poetry.dependencies]
python = "^3.11"
sentry-sdk = "^1.8.0"

[tool.poetry.dev-dependencies]
plainframework = "<1.0.0"
"plain.pytest" = "<1.0.0"
"plain.code" = "<1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
